Fix healthbars blocked by nonliving passengers. (#138)
Currently Neat does not render healthbars for mobs with passengers, but it also does not render healthbars for non-living entities. This creates the aforementioned edge case and render no healthbar for the mob-stack.
I've added an extra condition to the passenger check, now if the first passenger of a mob is non-living, then the healthbar of that mob will still render.